chore: 홈 페이지 스타일 변경 및 상세 페이지와의 className 충돌 해결
