+++ package.json
@@ -18,7 +18,8 @@
     "react-dom": "^19.0.0",
     "react-dropzone": "^14.3.5",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.2.0"
+    "react-router-dom": "^7.2.0",
+    "sweetalert2": "^11.17.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
