+++ src/App.css
@@ -4,11 +4,6 @@ body {
   padding: 0;
 }
 
-#root {
-  min-width: 100vw;
-  overflow-x: hidden;
-}
-
 @font-face {
   font-family: 'S-CoreDream-3Light';
   src: url('https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_six@1.2/S-CoreDream-3Light.woff') format('woff');
