+++ src/pages/Home.css
@@ -3,16 +3,48 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    justify-content: flex-start;  /* 헤더 아래로 정상 정렬 */
+    justify-content: flex-start; /* 헤더 아래로 정상 정렬 */
     width: 100%;
     height: 100vh;
-    padding-top: 60px;  /* 헤더 높이만큼 여백 추가 */
-    background-color: #f5f5f5;  /* 배경색 설정 */
+    background-color: #f5f5f5; /* 배경색 설정 */
+    position: relative;
+    background: #fff;
+    z-index: 1; /* 사이드바보다 아래 층 */
+    padding: 30px 0px;
   }
   
   .main-title {
-    font-size: 24px;
+    font-size: 34px;
     color: #0086CD;  /* 글씨 색 */
+    /* color: #47ADEB; */
+    font-weight: bold;
+    margin: 10px 0;
+  }
+  
+  .loading-spinner {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 400px;
+  }
+  
+  .spinner {
+    border: 4px solid #f3f3f3;
+    border-top: 4px solid #3498db;
+    border-radius: 50%;
+    width: 40px;
+    height: 40px;
+    animation: spin 1s linear infinite;
+  }
+  
+  @keyframes spin {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
   }
   
   .movies-grid {
@@ -21,9 +53,28 @@
     gap: 20px;
     justify-content: center;
     max-width: 1200px;
-    margin: auto;
+    margin-bottom: 3rem;
   }
   
+  /* 영화 목록 섹션 스타일 지정 */
+  .movies-section {
+    width: 100%;
+    margin: 1rem 3rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    text-align: center;
+  }
+
+  /* 영화 목록 섹션 제목 스타일 지정 */
+  .movies-section h2 {
+    margin: 20px;
+    display: block; 
+    color: #49454F;
+    font-size: 26px;
+  }
+
   .movie-card {
     background: #f9f9f9;
     padding: 15px;
@@ -39,12 +90,15 @@
   }
   
   .movie-title {
-    font-size: 18px;
-    font-weight: bold;
+    font-size: 1.2rem;
     margin-top: 10px;
+    padding-right: 1rem;
+    font-weight: bold;
   }
   
-  .movie-director, .movie-year, .movie-genre {
+  .movie-director,
+  .movie-year,
+  .movie-genre {
     font-size: 14px;
     margin: 5px 0;
   }
@@ -53,13 +107,13 @@
     display: inline-block;
     margin-top: 10px;
     color: white;
-    background: #0086CD;
+    background: #0086cd;
     padding: 5px 10px;
     border-radius: 5px;
     text-decoration: none;
   }
   
   .movie-link:hover {
-    background: #0066A2;
+    background: #0066a2;
   }
