--- src/pages/Main.css
@@ -1,96 +1,0 @@
-/* Main.css */
-.main-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start; /* 헤더 아래로 정상 정렬 */
-  width: 100%;
-  height: 100vh;
-  padding-top: 60px; /* 헤더 높이만큼 여백 추가 */
-  background-color: #f5f5f5; /* 배경색 설정 */
-  position: relative;
-  background: #fff;
-  z-index: 1; /* 사이드바보다 아래 층 */
-  padding: 20px;
-}
-
-.main-title {
-  font-size: 24px;
-  color: #0086cd; /* 글씨 색 */
-}
-
-.loading-spinner {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 400px;
-}
-
-.spinner {
-  border: 4px solid #f3f3f3;
-  border-top: 4px solid #3498db;
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-.movies-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-  gap: 20px;
-  justify-content: center;
-  max-width: 1200px;
-  margin: auto;
-}
-
-.movie-card {
-  background: #f9f9f9;
-  padding: 15px;
-  border-radius: 10px;
-  text-align: center;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.movie-poster {
-  width: 100%;
-  height: auto;
-  border-radius: 10px;
-}
-
-.movie-title {
-  font-size: 0.5em;
-  font-weight: bold;
-  margin-top: 10px;
-}
-
-.movie-director,
-.movie-year,
-.movie-genre {
-  font-size: 14px;
-  margin: 5px 0;
-}
-
-.movie-link {
-  display: inline-block;
-  margin-top: 10px;
-  color: white;
-  background: #0086cd;
-  padding: 5px 10px;
-  border-radius: 5px;
-  text-decoration: none;
-}
-
-.movie-link:hover {
-  background: #0066a2;
-}
